pyproject.toml
[project]
name = "rootcause-ai"
version = "0.1.0"
description = "AI-powered debugging assistant that analyzes codebases and explains root causes"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["ai", "debugging", "code-analysis", "langchain", "langgraph"]
authors = [
    { name = "RootCause AI Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-ollama>=0.1.0",
    "streamlit>=1.30.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
rootcause = "rootcause_ai.cli:main"

[project.gui-scripts]
rootcause-ui = "rootcause_ai.ui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rootcause_ai"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["rootcause_ai"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
